# licensing-service-client/src/main/resources/licensing-service-api-docs.yaml
openapi: 3.1.0
info:
  title: Licensing Service API
  description: Licensing Service API with type-safe generic responses using OpenAPI
  version: 2.0.0
servers:
  - url: http://localhost:8081/licensing-service
    description: Local service URL
security:
  - basicAuth: []
paths:
  /v1/licenses/access:
    post:
      tags:
        - license-controller
      operationId: createAccess
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueAccessRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ServiceResponseLicenseAccessResponse"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "405":
          description: Method Not Allowed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /v1/licenses/access/validate:
    post:
      tags:
        - license-controller
      operationId: validateAccess
      parameters:
        - name: License-Token
          in: header
          required: true
          schema:
            type: string
            maxLength: 400
            minLength: 200
            pattern: "^[A-Za-z0-9_-]+\\.[A-Za-z0-9_-]+\\.[A-Za-z0-9_-]+$"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateAccessRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ServiceResponseLicenseAccessResponse"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "405":
          description: Method Not Allowed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
components:
  schemas:
    IssueAccessRequest:
      type: object
      properties:
        licenseKey:
          type: string
          maxLength: 200
          minLength: 100
        instanceId:
          type: string
          maxLength: 100
          minLength: 10
        checksum:
          type: string
          maxLength: 200
          minLength: 40
        serviceId:
          type: string
          maxLength: 50
          minLength: 3
        serviceVersion:
          type: string
          maxLength: 20
          minLength: 3
        signature:
          type: string
          maxLength: 200
          minLength: 60
      required:
        - instanceId
        - licenseKey
        - serviceId
        - serviceVersion
        - signature
    LicenseAccessResponse:
      type: object
      properties:
        status:
          type: string
          enum:
            - TOKEN_CREATED
            - TOKEN_REFRESHED
            - TOKEN_ACTIVE
        licenseToken:
          type: string
    Meta:
      type: object
      properties:
        serverTime:
          type: string
          format: date-time
        sort:
          type: array
          items:
            $ref: "#/components/schemas/Sort"
    ServiceResponseLicenseAccessResponse:
      allOf:
        - $ref: "#/components/schemas/ServiceResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/LicenseAccessResponse"
      x-api-wrapper: true
      x-api-wrapper-datatype: LicenseAccessResponse
    Sort:
      type: object
      properties:
        field:
          type: string
        direction:
          type: string
          enum:
            - asc
            - desc
    ValidateAccessRequest:
      type: object
      properties:
        instanceId:
          type: string
          maxLength: 100
          minLength: 10
        checksum:
          type: string
          maxLength: 200
          minLength: 40
        serviceId:
          type: string
          maxLength: 50
          minLength: 3
        serviceVersion:
          type: string
          maxLength: 20
          minLength: 3
        signature:
          type: string
          maxLength: 200
          minLength: 60
      required:
        - instanceId
        - serviceId
        - serviceVersion
        - signature
    ErrorItem:
      type: object
      additionalProperties: false
      description: Standard error item structure.
      properties:
        code:
          type: string
          description: Short application-specific error code.
        message:
          type: string
          description: Human-readable error message.
        field:
          type: string
          description: Field name when error is field-specific.
        resource:
          type: string
          description: Domain resource name if applicable.
        id:
          type: string
          description: Resource identifier if applicable.
      required:
        - code
        - message
    ProblemDetail:
      type: object
      additionalProperties: false
      properties:
        type:
          type: string
          format: uri
          description: Problem type as a URI.
        title:
          type: string
          description: "Short, human-readable summary of the problem type."
        status:
          type: integer
          format: int32
          description: HTTP status code for this problem.
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri
          description: URI that identifies this specific occurrence.
        errorCode:
          type: string
          description: Application-specific error code.
        extensions:
          type: object
          additionalProperties: false
          description: Additional problem metadata.
          properties:
            errors:
              type: array
              description: List of error items (field-level or domain-specific).
              items:
                $ref: "#/components/schemas/ErrorItem"
    ServiceResponse:
      type: object
      properties:
        data:
          type: object
        meta:
          $ref: "#/components/schemas/Meta"
    ServiceResponseVoid:
      type: object
      properties:
        data:
          type: object
        meta:
          $ref: "#/components/schemas/Meta"
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic